+++ pHake/SDK/Weapon.hpp
@@ -13,7 +13,7 @@ public:
 		ammoinfo = AmmoInfo(proc);
 	}
 
-	void UpdateAll(uint64_t baseAddress)
+	void UpdateAll(uintptr_t baseAddress)
 	{
 		this->Update(baseAddress);
 		ammoinfo.Update(process_->read_multi_addr(this->base_, { 0x60, 0x8, 0x0 }));
